+++ topi/python/topi/cpp.py
@@ -1,4 +1,27 @@
 """FFI for C++ TOPI ops and schedules"""
+import sys
+import os
+import ctypes
 from tvm._ffi.function import _init_api_prefix
+from tvm._ffi import libinfo
+
+def _get_lib_names():
+    if sys.platform.startswith('win32'):
+        return ['libtvm_topi.dll', 'tvm_topi.dll']
+    if sys.platform.startswith('darwin'):
+        return ['libtvm_topi.dylib', 'tvm_topi.dylib']
+    return ['libtvm_topi.so', 'tvm_topi.so']
+
+def _load_lib():
+    """Load libary by searching possible path."""
+    curr_path = os.path.dirname(os.path.abspath(os.path.expanduser(__file__)))
+    lib_search = curr_path
+    lib_path = libinfo.find_lib_path(_get_lib_names(), lib_search, optional=True)
+    if lib_path is None:
+        return None, None
+    lib = ctypes.CDLL(lib_path[0], ctypes.RTLD_GLOBAL)
+    return lib, os.path.basename(lib_path[0])
+
+_LIB, _LIB_NAME = _load_lib()
 
 _init_api_prefix("topi.cpp", "topi")
